move RowCountToken to rewrite package
